First pass: move ToJWT/FromJWT to servant-auth
