+++ servant-auth-server/src/Servant/Auth/Server.hs
@@ -142,6 +142,7 @@ import Prelude hiding                           (readFile, writeFile)
 import Data.ByteString                          (ByteString, writeFile, readFile)
 import Data.Default.Class                       (Default (def))
 import Servant.Auth
+import Servant.Auth.JWT
 import Servant.Auth.Server.Internal             ()
 import Servant.Auth.Server.Internal.BasicAuth
 import Servant.Auth.Server.Internal.Class
