+++ servant-auth-server/src/Servant/Auth/Server/Internal/JWT.hs
@@ -16,32 +16,10 @@ import qualified Data.Text            as T
 import           Data.Time            (UTCTime)
 import           Network.Wai          (requestHeaders)
 
+import Servant.Auth.JWT               (FromJWT(..), ToJWT(..))
 import Servant.Auth.Server.Internal.ConfigTypes
 import Servant.Auth.Server.Internal.Types
 
--- This should probably also be from ClaimSet
---
--- | How to decode data from a JWT.
---
--- The default implementation assumes the data is stored in the unregistered
--- @dat@ claim, and uses the @FromJSON@ instance to decode value from there.
-class FromJWT a where
-  decodeJWT :: Jose.ClaimsSet -> Either T.Text a
-  default decodeJWT :: FromJSON a => Jose.ClaimsSet -> Either T.Text a
-  decodeJWT m = case HM.lookup "dat" (m ^. Jose.unregisteredClaims) of
-    Nothing -> Left "Missing 'dat' claim"
-    Just v  -> case fromJSON v of
-      Error e -> Left $ T.pack e
-      Success a -> Right a
-
--- | How to encode data from a JWT.
---
--- The default implementation stores data in the unregistered @dat@ claim, and
--- uses the type's @ToJSON@ instance to encode the data.
-class ToJWT a where
-  encodeJWT :: a -> Jose.ClaimsSet
-  default encodeJWT :: ToJSON a => a -> Jose.ClaimsSet
-  encodeJWT a = Jose.addClaim "dat" (toJSON a) Jose.emptyClaimsSet
 
 -- | A JWT @AuthCheck@. You likely won't need to use this directly unless you
 -- are protecting a @Raw@ endpoint.
