+++ servant-auth/servant-auth.cabal
@@ -33,8 +33,14 @@ library
   default-extensions: AutoDeriveTypeable ConstraintKinds DataKinds DefaultSignatures DeriveFoldable DeriveFunctor DeriveGeneric DeriveTraversable FlexibleContexts FlexibleInstances FunctionalDependencies GADTs KindSignatures MultiParamTypeClasses OverloadedStrings RankNTypes ScopedTypeVariables TypeFamilies TypeOperators
   ghc-options: -Wall
   build-depends:
-      base    >= 4.9  && < 4.14
-    , servant >= 0.15 && < 0.18
+      base                    >= 4.9      && < 4.14
+    , aeson                   >= 1.3.1.1  && < 1.5
+    , jose                    >= 0.7.0.0  && < 0.9
+    , lens                    >= 4.16.1   && < 4.19
+    , servant                 >= 0.15     && < 0.18
+    , text                    >= 1.2.3.0  && < 1.3
+    , unordered-containers    >= 0.2.9.0  && < 0.3
   exposed-modules:
       Servant.Auth
+      Servant.Auth.JWT
   default-language: Haskell2010
